Re-structure the project to follow conventions.
For Ruby projects (Rails projects are different) the lib/ folder is
where the production code should go.

If you are using RSpec there should be an spec/ folder where you put
everything in. If you are using one of the xUnit test frameworks
(Minitest, Test::Unit) this folder should be test/ instead.

The folder structure in the spec/ or (test/) folder should reflect the
structure in lib/.

Rails projects are different as their spec/ or test/ folder reflects the
structure of the app/ folder. In this case the common practice is
defining a spec/lib (or test/lib) and reflect the structure there.

The folders you define should mimic the namespaces that you are using.
In this particular case since you have a module A and a module B inside
of it you should have the folder structure shown in this commit.

Also the file names should reflect the name of the classes you are
defining. This is not really a requirement for things to work but it's a
common practice (specially in Rails projects).

The spec file for each class should have the same file name as the class
itself with the "_spec" suffix.