Switch to expect syntax.
The newest RSpec deprecated the "should" syntax in favor of the "expect"
syntax. This change removes the need of monkey patching the Object
class in order to make the #should method available in every object.

Since the deprecation warnings are pretty annoying I guess you are still
using a 2x version of RSpec. Give RSpec 3 a try :)