Update the script to run the specs
Given the include sentences used in the specs and given that RSpec loads
all the specs that have to be run, the version of the MyClass class in
the top level namespace will not be the expected one.

The include sentences are effectively monkey-patching the version of the
MyClass class in the top level namespace. Which one is available depends
on the order the specs are loaded by RSpec.

An easy way to test this behaviour is changing the names of the spec
files so their positions swap when ordering them by their path
alphabetically. This will cause the spec that was failing to pass and
the one that was passing to fail.