Use subject instead of described_class.new
You can use another useful alias in RSpec tests: subject. The value of
subject defaults to a new instance of the described class constructed by
calling .new on the described class without any parameters.

You can override it for your convenience if you need to:

describe MyClass do
  subject { MyClass.new(param1, param2) }
  let(:param1) { 'foo' }
  let(:param2) { 'bar' }

  #...
end

Those let sentences are a common way of extracting values to be reused
between specs. The first time param1 is called from the spec it will
execute the bloc and memoize the value so it's never run again.